Add files via upload
The following functions reworked in R Studio are were used to create a special object that stores a matrix and caches its inverse. The first function, makeCacheMatrix creates a special “matrix”
The CacheSolve function computes the inverse of the special “matrix” returned by makeCacheMatrix preceeding it. If the inverse has already been calculated (and the matrix has not changed), then cacheSolve  retrievesthe inverse from the cache.
Following we test the function with a matrix, compute the inverse, retrieve the inverse from the cache list, change the call matrix to the inverse, and then compute the inverse on that and return the original function.